Add files via upload
This project uses computer vision and deep machine learning techniques to predict dog breeds from images.
First, we identify dog facial key points for each image using a convolutional neural network. These key points
are then used to extract features via SIFT descriptors and color histograms. We then compare a variety of
classification algorithms, which use these features to predict the breed of the dog shown in the image. Our
best classifier is an SVM with a linear kernel and it predicts the correct dog breed on its first guess 52% of
the time; 90% of the time the correct dog breed is in the top 10 predictions.